Allow customization of the token font
There is still the bug where the token seems to be trying to go to its
superview to use its font (if the superview is a KSTextField), but
fails.  This commit doesn't fix that issue.

This change does not break the API, only exposes a new property on the
KSToken.