Add option to prevent blanking the text field on begin edit
This should not break existing behaviour.

The benefit of exposing this option is that people who use this library
can now control whether the tokenField gets blanked on being clicked.

(for example, when the user tries to move the cursor in the text field,
the default behavior was/is for the field to be blanked)